+++ backend/__init__.py
@@ -1,4 +1,4 @@
 """Package level information
 """
 
-__version__ = "3.1.0"
+__version__ = "3.2.0"